Fixed-return-type intrinsics return Long, not the magnitude rule (#196)
Closes #193.

`Len`, `InStr`, `InStrRev`, `LBound`, `UBound` and `VarType` returned
**Integer** whenever the answer fitted Int16. VB6 returns **Long** from
all of them regardless of magnitude.

## Cause

None of these functions chose a type. They built their result from a C#
`int`, and `Vb6Value(int)` applies a **magnitude rule** — anything
fitting Int16 is reported as an Integer. That rule is correct for an
arithmetic literal, whose type genuinely *does* follow its magnitude. It
is wrong for a function with a fixed declared return type.
`Vb6Value(long)` has no such rule, so the fix is to construct these as
Long.

## Measured the whole family first, and found two the issue didn't have

| `TypeName` in VB6 | intrinsics |
|---|---|
| **Long** | `Len`, `InStr`, `InStrRev`, `LBound`, `UBound`, `DateDiff`,
**`VarType`** |
| **Integer** | `Asc`, `Sgn`, `Year`, `Month`, `Day`, `Hour`, `Minute`,
`Second`, `Weekday`, **`DatePart`** |
| operand-dependent | `Int`, `Fix` |

- **`VarType` is Long**, despite every `vbXxx` code fitting an Integer —
and the comment beside it asserted Integer.
- **`DateDiff` is Long while `DatePart` is Integer.** Same family, same
argument shapes, different widths.

**It is not one rule**, which is the point: a blanket "widen the
integer-returning intrinsics" would have been just as wrong as leaving
them alone. All ten Integer cases are now tests so that stays true, and
`Int`/`Fix` are left deliberately alone — they preserve their operand's
subtype (`Int(3)` is Integer, `Int(3.5)` is Double) and must keep going
through the magnitude rule. That's a test too, as the control for the
whole change.

## Test churn, and why it's the honest kind

**21 existing assertions across 6 test classes** pinned the old Integer.
Every one encoded HexIDE's bug rather than VB6's behaviour, so they're
corrected rather than worked around.

That includes the #190 test that deliberately asserted the wrong value
with a comment pointing at this issue — **and the pointer did its job.**
The divergence was recorded at the exact spot someone would otherwise
have "corrected" the test to match the bug.

## Verification

20 new tests asserting through `TypeName` rather than by value, because
the value was never wrong — only the subtype. It doesn't stay invisible,
though: it feeds the arithmetic result-type ladder, so `Len("hi") + 1`
is a Long now, and that's a test as well.

Suite green at **1191**. All four mechanisms mutation-tested:

| mutation | tests failed |
|---|---|
| `Len` back to int | 3 |
| `InStr` back to int | 6 |
| `LBound`/`UBound` back to int | 14 |
| `VarType` back to int | 4 |



---------

+++ IDE/HexIDE.Runtime/Interpreter/VB6BuiltIns.Array.cs
@@ -30,7 +30,8 @@ public partial class VB6BuiltIns
         // indexing bounds[dimension-1] directly would throw an ArgumentOutOfRangeException that On Error can't trap.
         if (dimension < 1 || dimension > array.Rank)
             throw new VBRunTimeException(VBStandardError.SubscriptOutOfRange);
-        return new Vb6Value(upper ? array.UpperBound(dimension) : array.LowerBound(dimension));
+        // Long, not the magnitude rule — LBound/UBound have a fixed declared return type. See InStr. (#193)
+        return new Vb6Value((long)(upper ? array.UpperBound(dimension) : array.LowerBound(dimension)));
     }
 
     // Builds a 0-based 1-D array from items. Empty items -> bounds (0,-1), matching VB6's empty-array result.
